+++ packages/expo/src/winter/runtime.native.ts
@@ -17,6 +17,8 @@ install('TextEncoderStream', () => require('./TextDecoderStream').TextEncoderStr
 install('URL', () => require('./url').URL);
 // https://url.spec.whatwg.org/#urlsearchparams
 install('URLSearchParams', () => require('./url').URLSearchParams);
+// https://webidl.spec.whatwg.org/#idl-DOMException
+install('DOMException', () => require('./DOMException').DOMException);
 // https://streams.spec.whatwg.org/#rs
 // ReadableStream is injected by Metro as a global